backend/requirements.txt
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
decorator==4.4.2
filelock==3.18.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
fsspec==2025.3.2
gevent==25.5.1
google-api-core==2.24.1
google-api-python-client==2.169.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.69.1
gradio_client==1.10.2
greenlet==3.2.3
grpcio==1.71.0rc2
grpcio-status==1.63.0rc1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.5.3
Mako==1.3.10
MarkupSafe==3.0.2
more-itertools==10.6.0
moviepy==1.0.3
numpy==2.3.0
optionaldict==0.1.2
packaging==25.0
pdf2image==1.16.3
Pillow==9.5.0
playwright==1.50.0
proglog==0.1.12
prompt_toolkit==3.0.51
proto-plus==1.26.0
protobuf==4.25.6
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.1.1
PyJWT==2.8.0
PyMuPDF==1.26.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
requests==2.32.3
rsa==4.9
ruff==0.4.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sqlacodegen==3.0.0
SQLAlchemy==2.0.40
tqdm==4.67.1
typeguard==4.4.2
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
websockets==13.0
wechatpy==1.8.18
Werkzeug==3.1.3
xmltodict==0.14.2
zope.event==5.0
zope.interface==7.2
pypinyin==0.54.0